Move models and collections into a united module
