--- api/server/models/ufs.d.ts
@@ -1,17 +1,0 @@
-declare module 'api/models/ufs' {
-  interface Picture {
-    _id?: string;
-    complete?: boolean;
-    extension?: string;
-    name?: string;
-    progress?: number;
-    size?: number;
-    store?: string;
-    token?: string;
-    type?: string;
-    uploadedAt?: Date;
-    uploading?: boolean;
-    url?: string;
-    userId?: string;
-  }
-}
